daemon: Reload container on config update
If a config update is performed, reload the container
after the config has been rewritten to make the changes
effective, if the container is stopped. Otherwise,
mark the container as unsynced with the file system.
Trigger reload of config on container stop if out of sync.

+++ daemon/smartcard.c
@@ -1091,10 +1091,13 @@ smartcard_free(smartcard_t *smartcard)
 int
 smartcard_release_pairing(container_t *container)
 {
+	int ret = 0;
 	char *path = smartcard_token_paired_file_new(container);
-	int ret = unlink(path);
-	if (ret != 0) {
-		ERROR_ERRNO("Failed to remove file %s", path);
+	if (file_exists(path)) {
+		ret = unlink(path);
+		if (ret != 0) {
+			ERROR_ERRNO("Failed to remove file %s", path);
+		}
 	}
 	mem_free0(path);
 	return ret;
